I think this is a workable version
This does an ok job at inferring an INSDC RNA type for Rfam families. The
basic idea is to try some more simple reliable methods and then fallback
to less precise ones. In addition we avoid inferring types that are too
generic. This is because we trust Rfam so much that we don't want to use
a more generic type if it is possible to find a better one from another
data source.